+++ packages/talker_dio_logger/lib/talker_dio_logger.dart
@@ -1,0 +1,4 @@
+library talker_dio_logger;
+
+export 'talker_dio_logger_settings.dart';
+export 'talker_dio_logger_interceptor.dart';