+++ packages/talker_flutter/lib/src/models/flutter_talker_error.dart
@@ -4,12 +4,13 @@ import 'package:talker_flutter/talker_flutter.dart';
 /// Add - on class to default [TalkerError]
 /// with Flutter [Color] adding from [TalkerFlutterAdapterInterface]
 class FlutterTalkerError extends TalkerError
-    implements TalkerFlutterAdapterInterface {
+    implements FlutterTalkerDataInterface {
   FlutterTalkerError(Error error, {this.color}) : super(error);
 
   @override
   final Color? color;
 
   /// Used for [TalkerScreen] only
+  @override
   String generateFlutterTextMessage() => generateTextMessage();
 }