+++ packages/talker_flutter/lib/src/models/flutter_talker_exception.dart
@@ -4,12 +4,13 @@ import 'package:talker_flutter/talker_flutter.dart';
 /// Add - on class to default [TalkerException]
 /// with Flutter [Color] adding from [TalkerFlutterAdapterInterface]
 class FlutterTalkerException extends TalkerException
-    implements TalkerFlutterAdapterInterface {
+    implements FlutterTalkerDataInterface {
   FlutterTalkerException(Exception exception, {this.color}) : super(exception);
 
   @override
   final Color? color;
 
   /// Used for [TalkerScreen] only
+  @override
   String generateFlutterTextMessage() => generateTextMessage();
 }