+++ packages/talker_flutter/lib/src/src.dart
@@ -1,6 +1,6 @@
 export 'ui/talker_screen.dart';
 export 'ui/talker_listener.dart';
-export 'ui/talker_history_builder.dart';
+export 'ui/builders/builders.dart';
 export 'extensions/extensions.dart';
 export 'models/models.dart';
 export 'ui/theme/theme.dart';