+++ packages/talker_flutter/lib/src/ui/talker_screen.dart
@@ -1,11 +1,12 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:group_button/group_button.dart';
+import 'package:share_plus/share_plus.dart';
 import 'package:talker_flutter/src/controller/talker_screen_controller.dart';
 import 'package:talker_flutter/src/ui/ui.dart';
 import 'package:talker_flutter/talker_flutter.dart';
+
+import 'widgets/actions_bottom_sheet/actions_bottom_sheet.dart';
 
 /// UI view for output of all Talker logs and errors
 class TalkerScreen extends StatefulWidget {
@@ -14,11 +15,21 @@ class TalkerScreen extends StatefulWidget {
     required this.talker,
     this.appBarTitle = 'Flutter talker',
     this.theme = const TalkerScreenTheme(),
+    this.itemsBuilder,
   }) : super(key: key);
 
+  /// Talker implementation
   final TalkerInterface talker;
+
+  /// Theme for customize [TalkerScreen]
   final TalkerScreenTheme theme;
+
+  /// Screen [AppBar] title
   final String appBarTitle;
+
+  /// Optional Builder to customize
+  /// log items cards in list
+  final TalkerDataBuilder? itemsBuilder;
 
   @override
   State<TalkerScreen> createState() => _TalkerScreenState();
@@ -44,59 +55,20 @@ class _TalkerScreenState extends State<TalkerScreen> {
             actions: [
               SizedBox(
                 width: 40,
-                child: Transform(
-                  alignment: Alignment.center,
-                  transform: _controller.isLogOrderReversed
-                      ? Matrix4.rotationX(pi)
-                      : Matrix4.identity(),
-                  child: IconButton(
-                    padding: EdgeInsets.zero,
-                    iconSize: 28,
-                    onPressed: _controller.toggleLogOrder,
-                    icon: const Icon(
-                      Icons.swap_vert,
-                    ),
-                  ),
-                ),
-              ),
-              SizedBox(
-                width: 40,
                 child: IconButton(
                   padding: EdgeInsets.zero,
                   iconSize: 28,
-                  onPressed: _toggleLogsExpanded,
-                  icon: Icon(
-                    _controller.expandedLogs
-                        ? Icons.visibility_outlined
-                        : Icons.visibility_off_outlined,
-                  ),
-                ),
-              ),
-              SizedBox(
-                width: 40,
-                child: IconButton(
-                  padding: EdgeInsets.zero,
-                  iconSize: 28,
-                  onPressed: _cleanHistory,
-                  icon: const Icon(Icons.delete_outline),
-                ),
-              ),
-              SizedBox(
-                width: 40,
-                child: IconButton(
-                  padding: EdgeInsets.zero,
-                  iconSize: 24,
-                  onPressed: () => _copyAllLogs(context),
-                  icon: const Icon(Icons.copy),
-                ),
-              ),
-              SizedBox(
-                width: 40,
-                child: IconButton(
-                  padding: EdgeInsets.zero,
-                  iconSize: 28,
-                  onPressed: _showFilter,
+                  onPressed: () => _showFilter(context),
                   icon: const Icon(Icons.filter_alt_outlined),
+                ),
+              ),
+              SizedBox(
+                width: 40,
+                child: IconButton(
+                  padding: EdgeInsets.zero,
+                  iconSize: 28,
+                  onPressed: () => _showActionsBottomSheet(context),
+                  icon: const Icon(Icons.more_vert_rounded),
                 ),
               ),
             ],
@@ -113,6 +85,9 @@ class _TalkerScreenState extends State<TalkerScreen> {
                   final data = filtredElements[_controller.isLogOrderReversed
                       ? filtredElements.length - 1 - i
                       : i];
+                  if (widget.itemsBuilder != null) {
+                    return widget.itemsBuilder!.call(context, data);
+                  }
                   return TalkerDataCard(
                     data: data,
                     onTap: () => _copyTalkerDataItemText(data),
@@ -128,19 +103,67 @@ class _TalkerScreenState extends State<TalkerScreen> {
     );
   }
 
-  void _copyTalkerDataItemText(TalkerDataInterface data) {
-    Clipboard.setData(
-      ClipboardData(
-        text: data.generateTextMessage(),
-      ),
+  Future<void> _showActionsBottomSheet(BuildContext context) async {
+    await showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      backgroundColor: Colors.transparent,
+      builder: (context) {
+        return ActionsBottomSheet(
+          actions: [
+            BottomSheetAction(
+              onTap: _controller.toggleLogOrder,
+              title: 'Reverse logs',
+              icon: Icons.swap_vert,
+            ),
+            BottomSheetAction(
+              onTap: () => _copyAllLogs(context),
+              title: 'Copy all logs',
+              icon: Icons.copy,
+            ),
+            BottomSheetAction(
+              onTap: _toggleLogsExpanded,
+              title: _controller.expandedLogs ? 'Collapse logs' : 'Expand logs',
+              icon: _controller.expandedLogs
+                  ? Icons.visibility_outlined
+                  : Icons.visibility_off_outlined,
+            ),
+            BottomSheetAction(
+              onTap: _cleanHistory,
+              title: 'Clean history',
+              icon: Icons.delete_outline,
+            ),
+            BottomSheetAction(
+              onTap: _shareLogsInFile,
+              title: 'Share logs file',
+              icon: Icons.ios_share_outlined,
+            ),
+          ],
+          talkerScreenTheme: widget.theme,
+        );
+      },
     );
-    _showSnackBar(context, 'Log item is copied for console');
   }
 
-  void _showFilter() {
-    showModalBottomSheet(
+  Future<void> _shareLogsInFile() async {
+    final path = await _controller.saveLogsInFile(
+      widget.talker.history.text,
+    );
+    Share.shareFiles([path]);
+  }
+
+  void _copyTalkerDataItemText(TalkerDataInterface data) {
+    Clipboard.setData(
+      ClipboardData(text: data.generateTextMessage()),
+    );
+    _showSnackBar(context, 'Log item is copied in clipboard');
+  }
+
+  Future<void> _showFilter(BuildContext context) async {
+    await showModalBottomSheet(
       context: context,
       backgroundColor: Colors.transparent,
+      isScrollControlled: true,
       builder: (context) {
         return TalkerScreenFilter(
           controller: _controller,